requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
clickhouse-driver==0.2.9
mysql-connector-python==9.0.0
redis==5.0.8
bcrypt==4.2.0
pandas==2.2.3
scikit-learn==1.5.2
sqlparse==0.5.1
python-dotenv==1.0.1
google-generativeai==0.8.2
openai==1.51.0
python-jose==3.4.0
psycopg2-binary==2.9.9